@@ -1,11 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
+    <None Remove="planets.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="planets.json" />
   </ItemGroup>
 
   <ItemGroup>
