@@ -44,7 +44,7 @@ else
 @code {
 
     [Parameter]
-    public string Name { get; set; }
+    public string Name { get; set; } = string.Empty;
 
     private Planet CurrentPlanet { get; set; }
 
