@@ -13,7 +13,7 @@
             </NavLink>
         </li>
         <li class="nav-item px-3">
-            <NavLink class="nav-link" href="planets">
+            <NavLink class="nav-link" href="theplanets">
                 <span class="oi oi-globe" aria-hidden="true"></span> The Planets
             </NavLink>
         </li>
@@ -28,7 +28,7 @@
 @code {
     private bool collapseNavMenu = true;
 
-    private string NavMenuCssClass => collapseNavMenu ? "collapse" : null;
+    private string NavMenuCssClass => collapseNavMenu ? "collapse" : string.Empty;
 
     private void ToggleNavMenu()
     {
