@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Domain;
 
-namespace Domain
+public interface IPlanetsApi
 {
-    public interface IPlanetsApi
-    {
-        Task<Planet> GetPlanetAsync(string name);
-        Task<IEnumerable<Planet>> GetPlanetsAsync();
-    }
+    Task<Planet> GetPlanetAsync(string name);
+    Task<IEnumerable<Planet>> GetPlanetsAsync();
 }
